Changed the way the vat type id for the customer is determined.

+++ src/Invoice/Source.php
@@ -142,7 +142,7 @@ abstract class Source
      */
     public function getSign()
     {
-        return (float) ($this->getType() === static::CreditNote ? -1.0 : 1.0);
+        return $this->getType() === static::CreditNote ? -1.0 : 1.0;
     }
 
     /**
